MF-364: Move logic to new module
