+++ packages/esm-app-shell/src/types.ts
@@ -8,6 +8,7 @@ declare global {
     importMapOverrides: {
       getCurrentPageMap: () => Promise<ImportMap>;
     };
+    unresolvedBackendDeps: UnresolvedBackendDependencies;
   }
 }
 
@@ -41,4 +42,25 @@ export type ActivatorDefinition = Activator | RegExp | string;
 
 export interface ModuleResolver {
   (): System.Module | Promise<System.Module>;
+}
+
+export interface UnresolvedBackendDependencies {
+  modulesWithMissingBackendModules: MissingBackendModules[];
+  modulesWithWrongBackendModulesVersion: MissingBackendModules[];
+}
+
+export interface BackendModule {
+  uuid: string;
+  version: string;
+  requiredVersion?: string;
+}
+
+export interface MissingBackendModules {
+  moduleName: string;
+  backendModules: BackendModule[];
+}
+
+export interface Module {
+  moduleName: string;
+  backendDependencies: object;
 }
